autenticación y protección de endpoints

+++ backend/controllers/tareaController.js
@@ -1,9 +1,11 @@
 const asyncHandler = require('express-async-handler')
 
 const Tarea = require('../model/tareaModel')
+const User = require('../model/userModel')
+
 
 const getTareas = asyncHandler(async (req, res) => {
-    const tareas = await Tarea.find()
+    const tareas = await Tarea.find({ user: req.user.id })
     res.status(200).json(tareas)
 })
 
@@ -15,7 +17,8 @@ const setTarea = asyncHandler(async (req, res) => {
     }
 
     const tarea = await Tarea.create({
-        texto: req.body.texto
+        texto: req.body.texto,
+        user: req.user.id
     })
 
 
@@ -26,12 +29,18 @@ const updateTarea = asyncHandler(async (req, res) => {
 
     const tarea = await Tarea.findById(req.params.id)
 
-    if(!tarea) {
+    if (!tarea) {
         res.status(400)
         throw new Error('Tarea no encontrada')
     }
 
-    const updatedTarea = await Tarea.findByIdAndUpdate(req.params.id, req.body, {new:true})
+    //verificamos que el user de la tarea sea igual que esl user del token
+    if(tarea.user.toString() !== req.user.id) {
+        res.status(401)
+        throw new Error('Acceso no autorizado')
+    }
+
+    const updatedTarea = await Tarea.findByIdAndUpdate(req.params.id, req.body, { new: true })
 
     res.status(200).json(updatedTarea)
 })
@@ -40,9 +49,14 @@ const deleteTarea = asyncHandler(async (req, res) => {
 
     const tarea = await Tarea.findById(req.params.id)
 
-    if(!tarea) {
+    if (!tarea) {
         res.status(400)
         throw new Error('Tarea no encontrada')
+    }
+
+    if(tarea.user.toString() !== req.user.id) {
+        res.status(401)
+        throw new Error('Acceso no autorizado')
     }
 
     const deletedTarea = await Tarea.findByIdAndDelete(req.params.id)
